+++ contracts/token/FractionalERC20.sol
@@ -8,6 +8,6 @@ import "./ERC20.sol";
  */
 contract FractionalERC20 is ERC20 {
 
-  uint public decimals;
+  uint8 public decimals;
 
 }
